Feature/16 dbd to ram

+++ test.py
@@ -1,6 +1,5 @@
 import os
 import xml.dom.minidom as dom
-import sqlite3 as db
 from modules.xml_to_ram import XMLtoRAM
 from modules.ram_to_dbd import RAMtoDBD
 from modules.dbd_to_ram import DBDtoRAM
@@ -9,6 +8,9 @@ xml = dom.parse(os.path.join("xml/", "xml_test.xml"))
 ram = XMLtoRAM(xml).xml_to_ram()
 RAMtoDBD(ram, "test.db").ram_to_dbd()
 ram2 = DBDtoRAM("test.db").schema()
-os.remove("test.db")
+if os.path.exists("test.db"):
+    os.remove("test.db")
 
-print(ram2.name)
+for t in ram2.tables:
+    for f in t.fields:
+        print(f.name)
